refactor(server): architecture stabilization (Phase 2)
Move auth business logic into the service layer, dedupe guards, validate
config at boot, and fix a broken watchlist endpoint. Behavior-preserving.

- ARCH-1: move all business logic and direct PrismaService calls out of
  AuthController into AuthService (password-reset flow, profile, avatar,
  /me, preferences). Controller is now a thin delegation layer; cookie
  handling stays in the controller as a transport concern.
- ARCH-2: remove the duplicate JwtAuthGuard; standardize on the single
  JwtGuard across the review controller/module.
- SEC-9: validate environment variables at boot with Joi in ConfigModule
  (JWT_SECRET and DATABASE_URL required; allowUnknown:true so other env
  keys pass). Fails fast on misconfig instead of crashing at first use.
- QUAL-1: fix watchlist clear() passing req.user.sub (always undefined)
  instead of req.user.id.
- QUAL-5: correct npm scripts — start:staging sets NODE_ENV=staging;
  start:production/start:prod run the built dist instead of the TS dev
  server. Railway uses railway.json startCommand, so deploy is unaffected.
