+++ server/package.json
@@ -23,10 +23,10 @@
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "cross-env NODE_ENV=development nest start",
     "start:dev": "cross-env NODE_ENV=development nest start --watch",
-    "start:staging": "cross-env NODE_ENV=development nest start --watch",
-    "start:production": "cross-env NODE_ENV=production nest start",
+    "start:staging": "cross-env NODE_ENV=staging nest start --watch",
+    "start:production": "cross-env NODE_ENV=production node dist/main",
     "start:debug": "cross-env NODE_ENV=development nest start --debug --watch",
-    "start:prod": "node dist/main",
+    "start:prod": "cross-env NODE_ENV=production node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
     "test:watch": "jest --watch",
